// src/app/model/order.model.ts
export type Order = {
    id?: number;
    productId?: number;
    productName?: string;
    productCategory?: string;
    productPrice?: number;
    productDescription?: string;
    quantity?: number;
    totalPrice?: number;
    customerName?: string;
    customerEmail?: string;
    customerPhone?: string;
    customerAddress?: string;
  }